refactor(setup): split apply.rs (3.2k lines) into apply/{run,system,database,pf_conf,rcd,sudoers,root_password} (#191)
Pure move; mod.rs re-exports apply, generate_pf_conf, sudoers_content,
set_root_password_noninteractive so main.rs/tests.rs are unchanged.
Helpers crossed between modules are pub(super) with explicit imports;
include_str! paths adjusted for the deeper directory; the write_file
call-site guard test skips its own file by full path; path comments in
deploy.sh/run-all.sh/schemas.rs updated; #421 let_underscore guard now
lists the four apply modules that contain the pattern.
