+++ aifw-setup/src/apply/pf_conf.rs
@@ -1,0 +1,267 @@
+//! `generate_pf_conf` — the base pf.conf with the AiFw anchor hooks.
+
+use crate::config::{DefaultPolicy, SetupConfig};
+
+/// Generate a pf.conf based on the setup configuration
+pub fn generate_pf_conf(config: &SetupConfig) -> String {
+    let mut lines = Vec::new();
+
+    lines.push("# AiFw — Generated pf.conf".to_string());
+    lines.push(format!(
+        "# Generated by aifw-setup on {}",
+        chrono::Utc::now().to_rfc3339()
+    ));
+    lines.push(String::new());
+
+    // Macros
+    lines.push(format!("wan_if = \"{}\"", config.wan_interface));
+    if let Some(ref lan) = config.lan_interface {
+        lines.push(format!("lan_if = \"{lan}\""));
+    }
+    if let Some(ref ip) = config.lan_ip {
+        let host = ip.split('/').next().unwrap_or("192.168.1.1");
+        let prefix = ip.split('/').nth(1).unwrap_or("24");
+        // Derive network address from host IP (e.g. 192.168.1.1 -> 192.168.1.0)
+        let octets: Vec<&str> = host.split('.').collect();
+        let net = if octets.len() == 4 {
+            format!("{}.{}.{}.0", octets[0], octets[1], octets[2])
+        } else {
+            host.to_string()
+        };
+        lines.push(format!("lan_net = \"{net}/{prefix}\""));
+    }
+    lines.push(String::new());
+
+    // Tables
+    lines.push("# Tables for overload protection".to_string());
+    lines.push("table <bruteforce> persist".to_string());
+    lines.push("table <ai_blocked> persist".to_string());
+    lines.push(String::new());
+
+    // Options
+    lines.push("# Options".to_string());
+    lines.push("set skip on lo0".to_string());
+    lines.push("set skip on pfsync0".to_string());
+    lines.push("set block-policy drop".to_string());
+    lines.push("set optimization aggressive".to_string());
+    // floating (not if-bound) is required for pfsync state migration on
+    // failover — without it, replicated states bind to the original
+    // interface and don't match traffic on the surviving node's iface.
+    lines.push("set state-policy floating".to_string());
+    lines.push(String::new());
+
+    // Scrub
+    lines.push("# Normalization".to_string());
+    lines.push("scrub in all".to_string());
+    lines.push(String::new());
+
+    // NAT (must come before filter rules in pf)
+    if config.lan_interface.is_some() && config.nat_enabled {
+        lines.push("# NAT — LAN masquerade".to_string());
+        lines.push("nat on $wan_if from $lan_net to any -> ($wan_if)".to_string());
+        lines.push(String::new());
+    }
+
+    // AiFw managed anchors — NAT anchors must come before filter anchors
+    lines.push("# AiFw NAT anchors".to_string());
+    lines.push("nat-anchor \"aifw\"".to_string());
+    lines.push("nat-anchor \"aifw-nat\"".to_string());
+    lines.push("rdr-anchor \"aifw-nat\"".to_string());
+    lines.push("nat-anchor \"aifw-vpn\"".to_string());
+    lines.push(String::new());
+    // DHCP pass rules — must come BEFORE anchors because anchors contain
+    // "block quick" rules that would drop DHCP broadcasts first
+    if config.dhcp_enabled {
+        let lan = config.lan_interface.as_deref().unwrap_or("$lan_if");
+        lines.push("# DHCP server — allow broadcast requests and replies".to_string());
+        lines.push(format!("pass in quick on {lan} proto udp from 0.0.0.0 port 68 to 255.255.255.255 port 67 label \"dhcp-discover\""));
+        lines.push(format!(
+            "pass out quick on {lan} proto udp from any port 67 to any port 68 label \"dhcp-reply\""
+        ));
+        lines.push(String::new());
+    }
+
+    // ICMPv6 neighbor discovery — unlike ARP (layer 2, unfiltered), ND is
+    // IPv6 traffic pf filters, and the default block policy would eat
+    // neighbor solicitations, breaking ALL IPv6 (incl. NAT64, #531).
+    lines.push("# ICMPv6 neighbor discovery — required for IPv6 to function".to_string());
+    lines.push(
+        "pass quick inet6 proto icmp6 icmp6-type { routersol, routeradv, neighbrsol, neighbradv }"
+            .to_string(),
+    );
+    lines.push(String::new());
+
+    lines.push("# AiFw filter anchors".to_string());
+    // Multi-WAN anchors MUST be evaluated first so policy-routing decisions
+    // (route-to / rtable) are set before general filtering.
+    lines.push("anchor \"aifw-pbr\"".to_string());
+    lines.push("anchor \"aifw-mwan-leak\"".to_string());
+    lines.push("anchor \"aifw-mwan-reply\"".to_string());
+    lines.push("anchor \"aifw\"".to_string());
+    lines.push("anchor \"aifw-nat\"".to_string());
+    lines.push("anchor \"aifw-ratelimit\"".to_string());
+    lines.push("anchor \"aifw-vpn\"".to_string());
+    lines.push("anchor \"aifw-geoip\"".to_string());
+    lines.push("anchor \"aifw-tls\"".to_string());
+    lines.push("anchor \"aifw-ha\"".to_string());
+    lines.push(String::new());
+
+    // Default policy
+    lines.push("# Default policy".to_string());
+    match config.default_policy {
+        DefaultPolicy::Standard => {
+            lines.push("block in log all".to_string());
+            lines.push("pass out all keep state".to_string());
+        }
+        DefaultPolicy::Strict => {
+            lines.push("block log all".to_string());
+        }
+        DefaultPolicy::Permissive => {
+            lines.push("pass all keep state".to_string());
+        }
+    }
+    lines.push(String::new());
+
+    // Anti-spoof
+    lines.push("# Anti-spoofing".to_string());
+    lines.push("antispoof quick for $wan_if".to_string());
+    if config.lan_interface.is_some() {
+        lines.push("antispoof quick for $lan_if".to_string());
+    }
+    lines.push(String::new());
+
+    // Block overloaded IPs
+    lines.push("# Overload block rules".to_string());
+    lines.push("block in quick from <bruteforce>".to_string());
+    lines.push("block in quick from <ai_blocked>".to_string());
+    lines.push(String::new());
+
+    // Allow ICMP
+    lines.push("# ICMP".to_string());
+    lines.push("pass in quick inet proto icmp icmp-type echoreq keep state".to_string());
+    lines.push(String::new());
+
+    // Allow traffic from LAN subnet only
+    if config.lan_ip.is_some() {
+        lines.push("# LAN subnet — allow inbound".to_string());
+        lines.push("pass in quick from $lan_net keep state label \"local-subnet\"".to_string());
+    }
+    lines.push(String::new());
+
+    // Management access (SSH + API/UI). Scoped to the LAN interface when one
+    // is configured so the control plane is not exposed on the WAN by
+    // default (#560). A box with no LAN (single-NIC / WAN-only) keeps the
+    // any-interface rule — otherwise the operator would have no way in.
+    match config.lan_interface.as_deref() {
+        Some(lan) => {
+            lines.push("# SSH access (LAN only)".to_string());
+            lines.push(format!(
+                "pass in quick on {lan} proto tcp to any port 22 keep state label \"ssh\""
+            ));
+            lines.push(String::new());
+            lines.push("# AiFw API/UI access (LAN only)".to_string());
+            lines.push(format!(
+                "pass in quick on {lan} proto tcp to any port {} keep state label \"aifw-api\"",
+                config.api_port
+            ));
+            lines.push(String::new());
+        }
+        None => {
+            // WAN-only appliance: management must remain reachable on the
+            // single interface or the operator locks themselves out.
+            lines
+                .push("# SSH access (no LAN configured — reachable on all interfaces)".to_string());
+            lines.push(
+                "pass in quick proto tcp to any port 22 keep state label \"ssh\"".to_string(),
+            );
+            lines.push(String::new());
+            lines.push(
+                "# AiFw API/UI access (no LAN configured — reachable on all interfaces)"
+                    .to_string(),
+            );
+            lines.push(format!(
+                "pass in quick proto tcp to any port {} keep state label \"aifw-api\"",
+                config.api_port
+            ));
+            lines.push(String::new());
+        }
+    }
+
+    // LAN to WAN pass rule
+    if config.lan_interface.is_some() {
+        lines.push("# LAN to WAN".to_string());
+        lines.push("pass in on $lan_if from $lan_net keep state".to_string());
+        lines.push(String::new());
+
+        // DHCP on LAN (client sends from 0.0.0.0, must allow from any on LAN)
+        if config.dhcp_enabled {
+            lines.push("# DHCP on LAN".to_string());
+            lines.push(
+                "pass in quick on $lan_if proto udp from any to any port { 67, 68 } keep state"
+                    .to_string(),
+            );
+            lines.push(String::new());
+        }
+    }
+
+    // NOTE: we intentionally do NOT `load anchor "aifw" from ...` here.
+    //
+    // The aifw-daemon / aifw-api processes populate the `aifw` anchor directly
+    // via `pfctl -a aifw -f -` from the DB on startup and on every rule change.
+    // If pf.conf.aifw ALSO loaded the anchor from a stale on-disk file, any
+    // pf.conf reload (service pf reload, `aifw reconcile`, setup re-apply)
+    // would wipe the DB-driven rules back to whatever was in the file at
+    // install time — which is exactly how user edits to the LAN subnet rule
+    // silently regressed on reload. See root-cause write-up in v5.57.x notes.
+    //
+    // The anchor hook is already declared earlier via `anchor "aifw"`; it
+    // starts empty and gets populated by the daemon. The brief window
+    // between `service pf start` and `aifw_daemon start` is acceptable —
+    // the default-block catches inbound traffic, and daemon fills in the
+    // pass rules within ~1 second.
+    lines.push("# aifw anchor is populated at runtime by aifw-daemon from the DB.".to_string());
+    lines
+        .push("# Do NOT add `load anchor \"aifw\" from ...` here — see v5.57.3 notes.".to_string());
+    lines.push(String::new());
+
+    lines.join("\n")
+}
+
+/// Generate pf rules for testing (non-FreeBSD)
+#[cfg(test)]
+pub mod tests_support {
+    use super::*;
+    use crate::config::SetupConfig;
+
+    pub fn test_pf_conf() -> String {
+        let config = SetupConfig {
+            wan_interface: "em0".to_string(),
+            lan_interface: Some("em1".to_string()),
+            lan_ip: Some("192.168.1.1/24".to_string()),
+            api_port: 8080,
+            ..Default::default()
+        };
+        generate_pf_conf(&config)
+    }
+}
+
+#[cfg(test)]
+mod pf_conf_tests {
+    /// Guard (#531): the generated pf.conf must pass ICMPv6 neighbor
+    /// discovery before the anchors/default policy. ND is filterable IPv6
+    /// traffic (unlike ARP) — without this rule the default block policy
+    /// silently kills all IPv6, including NAT64.
+    #[test]
+    fn icmp6_neighbor_discovery_passes_before_anchors() {
+        let conf = super::tests_support::test_pf_conf();
+        let nd = conf
+            .find("proto icmp6 icmp6-type { routersol, routeradv, neighbrsol, neighbradv }")
+            .expect("ICMPv6 ND pass rule missing from generated pf.conf");
+        // Leading newline — a bare substring search would match the earlier
+        // `nat-anchor "aifw"` line instead of the filter anchor.
+        let anchors = conf
+            .find("\nanchor \"aifw\"")
+            .expect("filter anchors missing");
+        assert!(nd < anchors, "ND pass must precede the filter anchors");
+    }
+}
